fix: generic API
