+++ src/index.js
@@ -96,6 +96,7 @@ import FeatureFlagsController from './controllers/feature-flags.js';
 import routeRequiredCapabilities from './routes/required-capabilities.js';
 import ContactSalesLeadsController from './controllers/contact-sales-leads.js';
 import PageRelationshipsController from './controllers/page-relationships.js';
+import ToolsController from './controllers/tools.js';
 
 const uuidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
 
@@ -231,6 +232,7 @@ async function run(request, context) {
     const contactSalesLeadsController = ContactSalesLeadsController(context);
     const featureFlagsController = FeatureFlagsController(context);
     const pageRelationshipsController = PageRelationshipsController(context);
+    const toolsController = ToolsController(context, log, context.env);
 
     const routeHandlers = getRouteHandlers(
       auditsController,
@@ -280,6 +282,7 @@ async function run(request, context) {
       contactSalesLeadsController,
       featureFlagsController,
       pageRelationshipsController,
+      toolsController,
       ephemeralRunController,
     );
