+++ test/controllers/sites.test.js
@@ -139,7 +139,6 @@ describe('Sites Controller', () => {
     'getSiteMetricsBySource',
     'getPageMetricsBySource',
     'resolveSite',
-    'resolveConfig',
     'triggerBrandProfile',
     'getGraph',
   ];
@@ -5036,164 +5035,6 @@ describe('Sites Controller', () => {
 
       expect(response.status).to.equal(403);
       expect(error.message).to.equal('Restricted Operation');
-    });
-  });
-
-  describe('resolveConfig', () => {
-    it('should return 404 when site not found', async () => {
-      mockDataAccess.Site.findById.resolves(null);
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://github.com/adobe/test-repo' };
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(404);
-    });
-
-    it('should return 400 when gitHubURL is missing', async () => {
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = {};
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(400);
-      const body = await response.json();
-      expect(body.message).to.equal('gitHubURL is required');
-    });
-
-    it('should return 400 when gitHubURL is invalid', async () => {
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://not-a-github-url.com/foo' };
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(400);
-      const body = await response.json();
-      expect(body.message).to.equal('Invalid GitHub repository URL');
-    });
-
-    it('should return 500 when HLX_ADMIN_TOKEN is not configured', async () => {
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://github.com/adobe/test-repo' };
-      context.env.HLX_ADMIN_TOKEN = undefined;
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(500);
-    });
-
-    it('should return resolved hlxConfig and code on success', async () => {
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://github.com/adobe/test-repo' };
-      context.env.HLX_ADMIN_TOKEN = 'test-token';
-
-      // Mock the admin API call
-      nock('https://admin.hlx.page')
-        .get('/config/adobe/aggregated/test-repo.json')
-        .reply(200, {
-          cdn: { prod: { host: 'main--test-repo--adobe.aem.live' } },
-          code: { owner: 'adobe', repo: 'test-repo', source: 'https://github.com/adobe/test-repo' },
-          content: { source: { type: 'onedrive', url: 'https://adobe.sharepoint.com/sites/test' } },
-        });
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(200);
-
-      const body = await response.json();
-      expect(body.hlxConfig).to.be.an('object');
-      expect(body.hlxConfig.rso).to.deep.include({ owner: 'adobe', site: 'test-repo' });
-      expect(body.hlxConfig.cdn).to.be.an('object');
-      expect(body.hlxConfig.code).to.be.an('object');
-      expect(body.code).to.deep.include({
-        type: 'github',
-        owner: 'adobe',
-        repo: 'test-repo',
-        ref: 'main',
-        url: 'https://github.com/adobe/test-repo',
-      });
-    });
-
-    it('should fall back to fstab.yaml when admin API returns 404', async () => {
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://github.com/adobe/test-repo' };
-      context.env.HLX_ADMIN_TOKEN = 'test-token';
-
-      nock('https://admin.hlx.page')
-        .get('/config/adobe/aggregated/test-repo.json')
-        .reply(404);
-
-      nock('https://raw.githubusercontent.com')
-        .get('/adobe/test-repo/main/fstab.yaml')
-        .reply(200, 'mountpoints:\n  /: https://adobe.sharepoint.com/sites/test');
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(200);
-
-      const body = await response.json();
-      expect(body.hlxConfig.content.source.type).to.equal('onedrive');
-      expect(body.hlxConfig.content.source.url).to.equal('https://adobe.sharepoint.com/sites/test');
-      expect(body.code.owner).to.equal('adobe');
-      expect(body.code.repo).to.equal('test-repo');
-    });
-
-    it('should return basic hlxConfig when both admin API and fstab fail', async () => {
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://github.com/adobe/test-repo' };
-      context.env.HLX_ADMIN_TOKEN = 'test-token';
-
-      nock('https://admin.hlx.page')
-        .get('/config/adobe/aggregated/test-repo.json')
-        .reply(404);
-
-      nock('https://raw.githubusercontent.com')
-        .get('/adobe/test-repo/main/fstab.yaml')
-        .reply(404);
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(200);
-
-      const body = await response.json();
-      expect(body.hlxConfig.rso).to.deep.equal({ owner: 'adobe', site: 'test-repo', ref: 'main' });
-      expect(body.hlxConfig.content).to.be.undefined;
-      expect(body.code.type).to.equal('github');
-    });
-
-    it('should return 403 when user does not have access to the site', async () => {
-      sandbox.stub(AccessControlUtil.prototype, 'hasAccess').returns(false);
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://github.com/adobe/test-repo' };
-
-      const response = await sitesController.resolveConfig(context);
-      expect(response.status).to.equal(403);
-    });
-
-    it('should return 500 when resolveHlxConfigFromGitHubURL throws', async () => {
-      context.params = { siteId: SITE_IDS[0] };
-      context.data = { gitHubURL: 'https://github.com/adobe/test-repo' };
-      context.env.HLX_ADMIN_TOKEN = 'test-token';
-
-      // Force a network error on the admin API call
-      nock('https://admin.hlx.page')
-        .get('/config/adobe/aggregated/test-repo.json')
-        .replyWithError('ECONNREFUSED');
-
-      // Also force fstab to throw
-      nock('https://raw.githubusercontent.com')
-        .get('/adobe/test-repo/main/fstab.yaml')
-        .replyWithError('ECONNREFUSED');
-
-      // The function itself catches fstab errors, so we need to make the URL parsing fail
-      // Use esmock to override resolveHlxConfigFromGitHubURL to throw
-      const SitesControllerMocked = (await import('esmock')).default(
-        '../../src/controllers/sites.js',
-        {
-          '../../src/support/hlx-config.js': {
-            resolveHlxConfigFromGitHubURL: sinon.stub().rejects(new Error('Unexpected failure')),
-          },
-        },
-      );
-
-      const mockedController = (await SitesControllerMocked)(context, loggerStub, context.env);
-      const response = await mockedController.resolveConfig(context);
-      expect(response.status).to.equal(500);
-      const body = await response.json();
-      expect(body.message).to.equal('Failed to resolve config');
     });
   });
 });
