+++ test/routes/index.test.js
@@ -413,6 +413,10 @@ describe('getRouteHandlers', () => {
     search: sinon.stub(),
   };
 
+  const mockToolsController = {
+    resolveConfig: sinon.stub(),
+  };
+
   const mockEphemeralRunController = {
     batchRun: () => null,
     batchStatus: () => null,
@@ -467,6 +471,7 @@ describe('getRouteHandlers', () => {
       mockContactSalesLeadsController,
       mockFeatureFlagsController,
       mockPageRelationshipsController,
+      mockToolsController,
       mockEphemeralRunController,
     );
 
@@ -491,6 +496,7 @@ describe('getRouteHandlers', () => {
       'GET /trigger',
       'POST /event/fulfillment',
       'POST /slack/channels/invite-by-user-id',
+      'POST /tools/resolve-config',
       'POST /tools/api-keys',
       'GET /tools/api-keys',
       'POST /tools/import/jobs',
@@ -522,6 +528,7 @@ describe('getRouteHandlers', () => {
     expect(staticRoutes['GET /sites.csv']).to.equal(mockSitesController.getAllAsCsv);
     expect(staticRoutes['GET /sites.xlsx']).to.equal(mockSitesController.getAllAsExcel);
     expect(staticRoutes['GET /trigger']).to.equal(mockTrigger);
+    expect(staticRoutes['POST /tools/resolve-config']).to.equal(mockToolsController.resolveConfig);
     expect(staticRoutes['POST /tools/api-keys']).to.equal(mockApiKeyController.createApiKey);
     expect(staticRoutes['GET /tools/api-keys']).to.equal(mockApiKeyController.getApiKeys);
     expect(staticRoutes['POST /consent-banner']).to.equal(mockConsentBannerController.takeScreenshots);
@@ -734,7 +741,6 @@ describe('getRouteHandlers', () => {
       'GET /sites/:siteId/brand-guidelines',
       'GET /sites/:siteId/brand-profile',
       'POST /sites/:siteId/brand-profile',
-      'POST /sites/:siteId/resolve-config',
       'GET /sites/:siteId/page-citability/counts',
       'GET /sites/:siteId/top-pages',
       'GET /sites/:siteId/top-pages/:source',
